refactor(lotto): add the lotto repository for the generate lotto number
