+++ src/test/java/lotto/LottoTest.java
@@ -1,12 +1,9 @@
 package lotto;
 
-import camp.nextstep.edu.missionutils.Randoms;
+import lotto.domain.Lotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
 import java.util.List;
-
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class LottoTest {
@@ -23,16 +20,6 @@ class LottoTest {
         // TODO: 이 테스트가 통과할 수 있게 구현 코드 작성
         assertThatThrownBy(() -> new Lotto(List.of(1, 2, 3, 4, 5, 5)))
                 .isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @DisplayName("로또 번호생성 후 로또 번호 반환을 확인한다.")
-    @Test
-    void createLottoByRandomNumberAndGet() {
-        Lotto lotto = new Lotto(List.of(1, 2, 3, 4, 5, 6));
-        List<Integer> targetNumber = lotto.generateLotto();
-        lotto.setNumbers(targetNumber);
-
-        assertThat(targetNumber).isEqualTo(lotto.getNumbers());
     }
 
 }
